+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 * [AT-3314]
-  * Update AssignResources call in conftest.py to invoke central node instead of subarray node
+  * Update AssignResources and ReleaseResources calls in conftest.py to invoke central node instead of subarray node
 ## 27.1.0
 * [MAP-377]
   * Add OS shutdown command to gracefully shut down the talon boards prior to powering off the LRU