+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * [AT-3306]
   * Add new E2E test which does not call telescope OFF at the end
   * TelescopeON only called if telescope is not already ON
+* [AT-3314]
+  * Update AssignResources call in conftest.py to invoke central node instead of subarray node
 
 ## 27.1.0
 * [MAP-377]