+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 * [AT-3314]
   * Update AssignResources and ReleaseResources calls in conftest.py to invoke central node instead of subarray node
+* [AT-3311]
+  * Extend EDA config to all 4 ITF dishes
+
 ## 27.1.0
 * [MAP-377]
   * Add OS shutdown command to gracefully shut down the talon boards prior to powering off the LRU