+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 * [AT-3314]
-  * Update AssignResources call in conftest.py to invoke central node instead of subarray node
+  * Update AssignResources and ReleaseResources calls in conftest.py to invoke central node instead of subarray node
 * [AT-3311]
   * Extend EDA config to all 4 ITF dishes
+
 ## 27.1.0
 * [MAP-377]
   * Add OS shutdown command to gracefully shut down the talon boards prior to powering off the LRU