+++ CHANGELOG.md
@@ -5,6 +5,10 @@
   * Update AssignResources and ReleaseResources calls in conftest.py to invoke central node instead of subarray node
 * [AT-3311]
   * Extend EDA config to all 4 ITF dishes
+* [AT-3296]
+  * Update SDP to 1.2.3 and dish-lmc to 8.1.0
+    * This also required a vis-receive update to 5.4.3 (in submodule)
+    * As well as tango utils and tango base to upgrade to 0.4.22
 
 ## 27.1.0
 * [MAP-377]